+++ src/Juniper.Root/Units/Acceleration/MetersPerSecondSquared.cs
@@ -8,19 +8,19 @@ namespace Juniper.Units
         /// <summary>
         /// Gravitational acceleration constant
         /// </summary>
-        public const float GRAVITY = 9.80665f;
+        public const double GRAVITY = 9.80665;
 
         /// <summary>
         /// Conversion factor from feet per second squared to meteers per second squared.
         /// </summary>
-        public const float PER_FOOT_PER_SECOND_SQUARED = Units.Meters.PER_FOOT;
+        public const double PER_FOOT_PER_SECOND_SQUARED = Units.Meters.PER_FOOT;
 
         /// <summary>
         /// Convert from meters per second squared to feet per second squared.
         /// </summary>
         /// <param name="mpsps">The number of meters per second squared</param>
         /// <returns>The number of feet per second squared</returns>
-        public static float FeetPerSecondSquared(float mpsps)
+        public static double FeetPerSecondSquared(double mpsps)
         {
             return mpsps * Units.FeetPerSecondSquared.PER_METER_PER_SECOND_SQUARED;
         }
