+++ src/Juniper.Root/Units/Area/Acres.cs
@@ -8,59 +8,59 @@ namespace Juniper.Units
         /// <summary>
         /// Conversion factor from square micrometers to acres.
         /// </summary>
-        public const float PER_SQUARE_MICROMETER = 1 / Units.SquareMicrometers.PER_ACRE;
+        public const double PER_SQUARE_MICROMETER = 1 / Units.SquareMicrometers.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from square millimeters to acres.
         /// </summary>
-        public const float PER_SQUARE_MILLIMETER = 1 / Units.SquareMillimeters.PER_ACRE;
+        public const double PER_SQUARE_MILLIMETER = 1 / Units.SquareMillimeters.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from square centimeters to acres.
         /// </summary>
-        public const float PER_SQUARE_CENTIMETER = 1 / Units.SquareCentimeters.PER_ACRE;
+        public const double PER_SQUARE_CENTIMETER = 1 / Units.SquareCentimeters.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from square inches to acres.
         /// </summary>
-        public const float PER_SQUARE_INCH = 1 / Units.SquareInches.PER_ACRE;
+        public const double PER_SQUARE_INCH = 1 / Units.SquareInches.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from acres to acres.
         /// </summary>
-        public const float PER_SQUARE_FOOT = 1 / Units.SquareFeet.PER_ACRE;
+        public const double PER_SQUARE_FOOT = 1 / Units.SquareFeet.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from square rod to acres.
         /// </summary>
-        public const float PER_SQUARE_ROD = 1 / Units.SquareRods.PER_ACRE;
+        public const double PER_SQUARE_ROD = 1 / Units.SquareRods.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from acres to acres.
         /// </summary>
-        public const float PER_SQUARE_YARD = 1 / Units.SquareYards.PER_ACRE;
+        public const double PER_SQUARE_YARD = 1 / Units.SquareYards.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from square meters to acres.
         /// </summary>
-        public const float PER_SQUARE_METER = 1 / Units.SquareMeters.PER_ACRE;
+        public const double PER_SQUARE_METER = 1 / Units.SquareMeters.PER_ACRE;
 
         /// <summary>
         /// Conversion factor from square kilometers to acres.
         /// </summary>
-        public const float PER_SQUARE_KILOMETER = PER_SQUARE_METER * Units.SquareMeters.PER_SQUARE_KILOMETER;
+        public const double PER_SQUARE_KILOMETER = PER_SQUARE_METER * Units.SquareMeters.PER_SQUARE_KILOMETER;
 
         /// <summary>
         /// Conversion factor from square miles to acres.
         /// </summary>
-        public const float PER_SQUARE_MILE = 640;
+        public const double PER_SQUARE_MILE = 640;
 
         /// <summary>
         /// Convert from acres to square micrometers.
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square micrometers</returns>
-        public static float SquareMicrometers(float acres)
+        public static double SquareMicrometers(double acres)
         {
             return acres * Units.SquareMicrometers.PER_ACRE;
         }
@@ -70,7 +70,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square millimeters</returns>
-        public static float SquareMillimeters(float acres)
+        public static double SquareMillimeters(double acres)
         {
             return acres * Units.SquareMillimeters.PER_ACRE;
         }
@@ -80,7 +80,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square centimeters</returns>
-        public static float SquareCentimeters(float acres)
+        public static double SquareCentimeters(double acres)
         {
             return acres * Units.SquareCentimeters.PER_ACRE;
         }
@@ -90,7 +90,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square inches</returns>
-        public static float SquareInches(float acres)
+        public static double SquareInches(double acres)
         {
             return acres * Units.SquareInches.PER_ACRE;
         }
@@ -100,7 +100,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of acres</returns>
-        public static float SquareFeet(float acres)
+        public static double SquareFeet(double acres)
         {
             return acres * Units.SquareFeet.PER_ACRE;
         }
@@ -110,7 +110,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of acres</returns>
-        public static float SquareYards(float acres)
+        public static double SquareYards(double acres)
         {
             return acres * Units.SquareYards.PER_ACRE;
         }
@@ -120,7 +120,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square centimeters</returns>
-        public static float SquareMeters(float acres)
+        public static double SquareMeters(double acres)
         {
             return acres * Units.SquareMeters.PER_ACRE;
         }
@@ -130,7 +130,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square rods</returns>
-        public static float SquareRods(float acres)
+        public static double SquareRods(double acres)
         {
             return acres * Units.SquareRods.PER_ACRE;
         }
@@ -140,7 +140,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square kilometers</returns>
-        public static float SquareKilometers(float acres)
+        public static double SquareKilometers(double acres)
         {
             return acres * Units.SquareKilometers.PER_ACRE;
         }
@@ -150,7 +150,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="acres">The number of acres</param>
         /// <returns>The number of square miles</returns>
-        public static float SquareMiles(float acres)
+        public static double SquareMiles(double acres)
         {
             return acres * Units.SquareMiles.PER_ACRE;
         }
