+++ src/Juniper.Root/Units/Light/Brightness.cs
@@ -8,19 +8,19 @@ namespace Juniper.Units
         /// <summary>
         /// Conversion factor from Lumens to Brightness.
         /// </summary>
-        public const float PER_LUMEN = 1 / Units.Lumens.PER_BRIGHTNESS;
+        public const double PER_LUMEN = 1 / Units.Lumens.PER_BRIGHTNESS;
 
         /// <summary>
         /// Conversion factor from Nits to Brightness.
         /// </summary>
-        public const float PER_NIT = Units.Lumens.PER_NIT / Units.Lumens.PER_BRIGHTNESS;
+        public const double PER_NIT = Units.Lumens.PER_NIT / Units.Lumens.PER_BRIGHTNESS;
 
         /// <summary>
         /// Convert from Brightness to Lumens
         /// </summary>
         /// <param name="brightness">The number of brightness</param>
         /// <returns>The number of lumens</returns>
-        public static float Lumens(float brightness)
+        public static double Lumens(double brightness)
         {
             return brightness * Units.Lumens.PER_BRIGHTNESS;
         }
@@ -30,7 +30,7 @@ namespace Juniper.Units
         /// </summary>
         /// <param name="brightness">The number of brightness</param>
         /// <returns>The number of nits</returns>
-        public static float Nits(float brightness)
+        public static double Nits(double brightness)
         {
             return brightness * Units.Nits.PER_BRIGHTNESS;
         }
